Move vectorset writer API to a new module and handle delete conflict
